+++ jabgui/src/main/java/org/jabref/gui/openoffice/AdvancedCiteDialogViewModel.java
@@ -1,25 +1,23 @@
 package org.jabref.gui.openoffice;
 
-import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
+
+import org.jabref.model.openoffice.style.CitationType;
 
 public class AdvancedCiteDialogViewModel {
 
     private final StringProperty pageInfo = new SimpleStringProperty("");
-    private final BooleanProperty citeInPar = new SimpleBooleanProperty();
-    private final BooleanProperty citeInText = new SimpleBooleanProperty();
+    private final ObjectProperty<CitationType> citationType =
+            new SimpleObjectProperty<>(CitationType.AUTHORYEAR_INTEXT);
 
     public StringProperty pageInfoProperty() {
         return pageInfo;
     }
 
-    public BooleanProperty citeInParProperty() {
-        return citeInPar;
-    }
-
-    public BooleanProperty citeInTextProperty() {
-        return citeInText;
+    public ObjectProperty<CitationType> citationTypeProperty() {
+        return citationType;
     }
 }